Refactor and clean up achievements on user profile
*Cleaner and less custom styling
*Resembles the other content more
*No more card
*No more console warning
*More responsive on different viewports
*No overflow on mobile
*No magic numbers
*Improved naming
